wip: compile-time version of a check for enum items
While this check already revealed some mistakes, however, compilation
fails if the limits are out of our control (e.g. from a system header
like pfkeyv2.h), which might not be ideal as it could require a regular
user to patch the files.  So maybe we could hide this behind a
compile option?